requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aiohttp
websockets